+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Fixes
+
+- Fallback Uri parsing to `unknown` if its invalid ([#1395](https://github.com/getsentry/sentry-dart/pull/1395))
+
 ## 7.5.0
 
 ### Features